fix: Show map skeleton correctly
