+++ app/web/src/newhotness/Workspace.vue
@@ -640,6 +640,59 @@ body.light .scrollable {
   }
 }
 
+/* ================================================================================================
+ * Shimmer for skeleton divs
+ * ================================================================================================
+ */
+.skeleton-shimmer {
+  position: relative;
+  overflow: hidden;
+}
+
+.skeleton-shimmer::before {
+  content: "";
+  position: absolute;
+  top: 0;
+  left: -100%;
+  width: 100%;
+  height: 100%;
+  animation: shimmer 1.5s infinite;
+}
+
+/* Light theme shimmer */
+body.light .skeleton-shimmer::before {
+  background: linear-gradient(
+    90deg,
+    transparent,
+    rgba(255, 255, 255, 0.8),
+    transparent
+  );
+}
+
+/* Dark theme shimmer */
+body.dark .skeleton-shimmer::before {
+  background: linear-gradient(
+    90deg,
+    transparent,
+    rgba(255, 255, 255, 0.1),
+    transparent
+  );
+}
+
+@keyframes shimmer {
+  0% {
+    left: -100%;
+  }
+  100% {
+    left: 100%;
+  }
+}
+
+/* ================================================================================================
+ * END Shimmer for skeleton divs
+ * ================================================================================================
+ */
+
 // other cards may look differently, can be defined globally, or piecemeal
 
 /* TODO(Wendy) - temporary color classes for the new UI */
