+++ app/web/src/newhotness/layout_components/DelayedComponent.vue
@@ -1,0 +1,17 @@
+<!-- eslint-disable vue/no-multiple-template-root -->
+<template>
+  <slot v-if="show" />
+</template>
+
+<script setup lang="ts">
+import { onMounted, ref } from "vue";
+
+const DELAY_MS = 300;
+
+const show = ref(false);
+onMounted(() => {
+  setTimeout(() => {
+    show.value = true;
+  }, DELAY_MS);
+});
+</script>
