+++ app/web/src/newhotness/skeletons/ExploreGridSkeleton.vue
@@ -1,179 +1,20 @@
 <template>
-  <section
-    :class="clsx('grid h-full explore', themeClasses('bg-white', 'bg-black'))"
-  >
+  <DelayedComponent>
     <div
-      :class="
-        clsx(
-          'main pt-xs flex flex-col gap-xs items-stretch [&>div]:mx-[12px]',
-          themeClasses('bg-white', 'bg-black'),
-        )
-      "
+      :class="clsx('w-full grid gap-4 flex-1 pr-6')"
+      :style="{
+        'grid-template-columns': `repeat(${columnsCount}, minmax(0, 1fr))`,
+      }"
     >
-      <!-- view dropdown / search bar -->
-      <div class="flex-none flex flex-row items-center gap-xs">
-        <!-- View dropdown skeleton -->
-        <div
-          :class="
-            clsx(
-              'rounded min-w-[128px] h-8 skeleton-shimmer',
-              themeClasses(
-                'bg-neutral-200 border border-neutral-400',
-                'bg-neutral-700 border border-neutral-600',
-              ),
-            )
-          "
-        ></div>
-
-        <!-- search bar -->
-        <div
-          :class="
-            clsx(
-              'rounded grow h-8 skeleton-shimmer flex items-center px-2 gap-2',
-              themeClasses(
-                'bg-neutral-200 border border-neutral-400',
-                'bg-neutral-700 border border-neutral-600',
-              ),
-            )
-          "
-        ></div>
-      </div>
-
-      <div class="flex-none flex flex-row items-center gap-xs justify-between">
-        <!-- grid/map toggle -->
-        <div class="flex flex-row gap-1">
-          <div
-            :class="
-              clsx(
-                'px-3 py-1 rounded h-8 skeleton-shimmer min-w-[128px]',
-                themeClasses('bg-neutral-200', 'bg-neutral-700'),
-              )
-            "
-          ></div>
-        </div>
-
-        <div class="flex flex-row gap-xs">
-          <!-- group by dropdown skeleton -->
-          <div
-            :class="
-              clsx(
-                'rounded min-w-[100px] h-8 skeleton-shimmer',
-                themeClasses(
-                  'bg-neutral-200 border border-neutral-400',
-                  'bg-neutral-700 border border-neutral-600',
-                ),
-              )
-            "
-          ></div>
-
-          <!-- sort by dropdown skeleton -->
-          <div
-            :class="
-              clsx(
-                'rounded min-w-[100px] h-8 skeleton-shimmer',
-                themeClasses(
-                  'bg-neutral-200 border border-neutral-400',
-                  'bg-neutral-700 border border-neutral-600',
-                ),
-              )
-            "
-          ></div>
-        </div>
-      </div>
-
-      <!-- component grid skeleton -->
-      <div
-        :class="
-          clsx(
-            'w-full grid gap-4 flex-1 pr-6',
-            themeClasses('bg-white', 'bg-black'),
-          )
-        "
-        :style="{
-          'grid-template-columns': `repeat(${columnsCount}, minmax(0, 1fr))`,
-        }"
-      >
-        <GridTileSkeleton v-for="n in skeletonCount" :key="n" />
-      </div>
+      <GridTileSkeleton v-for="n in skeletonCount" :key="n" />
     </div>
-
-    <!-- Right panel -->
-    <div
-      :class="
-        clsx(
-          'right flex flex-col border-l',
-          themeClasses(
-            'bg-neutral-100 border-neutral-400',
-            'bg-neutral-800 border-neutral-600',
-          ),
-        )
-      "
-    >
-      <div class="grow grid grid-rows-2">
-        <!-- history skel -->
-        <div class="overflow-hidden flex flex-col">
-          <h3
-            :class="
-              clsx(
-                'flex flex-row items-center',
-                'sticky top-0 text-lg font-bold px-xs py-2',
-                'skeleton-shimmer',
-                themeClasses('bg-neutral-200', 'bg-neutral-900'),
-              )
-            "
-          >
-            <div
-              class="w-4 h-4 mr-2 bg-neutral-300 dark:bg-neutral-700 rounded skeleton-shimmer"
-            ></div>
-            <div
-              class="w-16 h-4 bg-neutral-300 dark:bg-neutral-700 rounded skeleton-shimmer"
-            ></div>
-          </h3>
-          <div class="flex-1 p-xs space-y-1">
-            <div
-              v-for="n in 8"
-              :key="n"
-              class="h-8 bg-neutral-200 dark:bg-neutral-700 rounded skeleton-shimmer"
-            ></div>
-          </div>
-        </div>
-
-        <!-- actions skel -->
-        <div class="overflow-hidden flex flex-col">
-          <h3
-            :class="
-              clsx(
-                'flex flex-row items-center',
-                'sticky top-0 text-lg font-bold px-xs py-2',
-                'skeleton-shimmer',
-                themeClasses('bg-neutral-200', 'bg-neutral-900'),
-              )
-            "
-          >
-            <div
-              class="w-4 h-4 mr-2 bg-neutral-300 dark:bg-neutral-700 rounded skeleton-shimmer"
-            ></div>
-            <div
-              class="w-16 h-4 bg-neutral-300 dark:bg-neutral-700 rounded skeleton-shimmer"
-            ></div>
-          </h3>
-          <div class="flex-1 p-xs space-y-1">
-            <div
-              v-for="n in 8"
-              :key="n"
-              class="h-8 bg-neutral-200 dark:bg-neutral-700 rounded skeleton-shimmer"
-            ></div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </section>
+  </DelayedComponent>
 </template>
 
 <script lang="ts" setup>
 import { computed } from "vue";
 import clsx from "clsx";
-import { themeClasses } from "@si/vue-lib/design-system";
+import DelayedComponent from "@/newhotness/layout_components/DelayedComponent.vue";
 import GridTileSkeleton from "./GridTileSkeleton.vue";
 import {
   windowWidthReactive,
@@ -213,19 +54,3 @@ const rowsCount = computed(() => {
 
 const skeletonCount = computed(() => columnsCount.value * rowsCount.value);
 </script>
-
-<style lang="css" scoped>
-section.grid.explore {
-  grid-template-columns: minmax(0, 70%) minmax(0, 30%);
-  grid-template-rows: 100%;
-  grid-template-areas: "main right";
-}
-
-div.main {
-  grid-area: main;
-}
-
-div.right {
-  grid-area: right;
-}
-</style>
