+++ app/web/src/newhotness/types.ts
@@ -28,6 +28,7 @@ export function assertIsDefined<T>(value: T | undefined): asserts value is T {
 export interface ExploreContext {
   viewId: ComputedRef<string>;
   upgradeableComponents: ComputedRef<Set<string>>;
+  showSkeleton: ComputedRef<boolean>;
 }
 
 // Define an enum for function kinds
